tools/autotile-converter/pyproject.toml
[project]
name = "autotile-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "typed-argument-parser>=1.11.0",
]
